исправлена критическая уязвимость. прри запуске тестов очищалась пользовательская база данных. теперь есть две отделные базы - для тестов и для хранения анкет пользователей
